Create Heine-Borel (incomplete)
The first two sections are a bit of auxiliary logic and some work on intervals. Then I proceed to (almost) prove that a closed interval is compact in the section "closed_intervals_are_compact". The final section, bounded_of_compact, is my incomplete attempt to prove that any compact set in the reals is bounded, the idea is to prove by contradiction: I create a covering together with a sequence of elements of the (supposedly) unbounded set so that any subcovering is infinite.